Graceful exit for queries in empty projects

+++ scripts/_distances_noref_nomulti.bash
@@ -101,23 +101,25 @@ if [[ "$CLASSIF" != "." ]] ; then
   fi
 fi
 
-# Build tree with medoids
-echo "select seq2 from $METRIC;" | sqlite3 "${DATASET}.${METRIC}.db" \
-  | sort | uniq > "${DATASET}.tmp0"
-perl -pe "s/^/^/" "${DATASET}.tmp0" | perl -pe "s/$/\\t/" \
-  > "${DATASET}.tmp1"
-perl -pe "s/^/\\t/" "${DATASET}.tmp0" | perl -pe "s/$/\\t/" \
-  > "${DATASET}.tmp2"
-echo "a b value" | tr " " "\\t" > "${DATASET}.txt"
-gzip -c -d "$REF_TABLE" | cut -f 2-4 \
-  | grep -f "${DATASET}.tmp1" | grep -f "${DATASET}.tmp2" \
-  >> "${DATASET}.txt"
-echo "select seq1, seq2, $METRIC from $METRIC;" \
-  | sqlite3 "${DATASET}.${METRIC}.db"  | tr "\\|" "\\t" \
-  >> "${DATASET}.txt"
-"$MIGA/utils/ref-tree.R" "${DATASET}.txt" "$DATASET" "$DATASET"
-rm "$DATASET".tmp[012] "${DATASET}.txt"
-
 # Finalize
 N=11
 miga-checkpoint_n
+
+# Build tree with medoids
+if [[ -s "${DATASET}.${METRIC}.db" ]] ; then
+  echo "select seq2 from $METRIC;" | sqlite3 "${DATASET}.${METRIC}.db" \
+    | sort | uniq > "${DATASET}.tmp0"
+  perl -pe "s/^/^/" "${DATASET}.tmp0" | perl -pe "s/$/\\t/" \
+    > "${DATASET}.tmp1"
+  perl -pe "s/^/\\t/" "${DATASET}.tmp0" | perl -pe "s/$/\\t/" \
+    > "${DATASET}.tmp2"
+  echo "a b value" | tr " " "\\t" > "${DATASET}.txt"
+  gzip -c -d "$REF_TABLE" | cut -f 2-4 \
+    | grep -f "${DATASET}.tmp1" | grep -f "${DATASET}.tmp2" \
+    >> "${DATASET}.txt"
+  echo "select seq1, seq2, $METRIC from $METRIC;" \
+    | sqlite3 "${DATASET}.${METRIC}.db"  | tr "\\|" "\\t" \
+    >> "${DATASET}.txt"
+  "$MIGA/utils/ref-tree.R" "${DATASET}.txt" "$DATASET" "$DATASET"
+  rm "$DATASET".tmp[012] "${DATASET}.txt"
+fi
